refactor: move descriptor creation into ooapi
